Social media video download script
If it is possible to use yt-dlp to find a video associated with a social network link (only open profiles, public channels and similar - the extraction of content marked as for over 18s, with content restrictions such as nationality, for example, will fail). The script performs the transcoding process preferably using Nvidia resources. It leaves the videos in x264, a format that WhatsApp accepts - It also breaks them into smaller files, adhering to WhatsApp's size limits.